(#17) feat: DirectionGap 방향 별로 분해
